修改

+++ librtp/payload/rtp-h264-unpack.c
@@ -2,7 +2,6 @@
 #include <memory.h>
 #include <assert.h>
 #include <errno.h>
-#include "cstringext.h"
 #include "ctypedef.h"
 #include "rtp-unpack.h"
 #include "rtp-packet.h"
@@ -49,7 +48,7 @@ static void rtp_h264_unpack_destroy(void* p)
 	free(unpacker);
 }
 
-static int rtp_h264_unpack_stap_a(struct rtp_h264_unpack_t *unpacker, const void* data, int bytes, uint64_t time)
+static int rtp_h264_unpack_stap_a(struct rtp_h264_unpack_t *unpacker, const void* data, size_t bytes, uint64_t time)
 {
 	// 5.7.1. Single-Time Aggregation Packet (STAP) (p20)
 	unsigned char stapnal;
@@ -82,7 +81,7 @@ static int rtp_h264_unpack_stap_a(struct rtp_h264_unpack_t *unpacker, const void
 	return 0;
 }
 
-static int rtp_h264_unpack_stap_b(struct rtp_h264_unpack_t *unpacker, const void* data, int bytes, uint64_t time)
+static int rtp_h264_unpack_stap_b(struct rtp_h264_unpack_t *unpacker, const void* data, size_t bytes, uint64_t time)
 {
 	// 5.7.1. Single-Time Aggregation Packet (STAP)
 	unsigned char stapnal;
